+++ R/rmw_do_all.R
@@ -86,10 +86,7 @@ rmw_do_all <- function(df, variables, variables_sample = NA, n_trees = 300,
   
   # Check inputs
   if (se && !keep_inbag) {
-    stop(
-      "To calculate standard errors, `keep_inbag` needs to be `TRUE`...", 
-      call. = FALSE
-    )
+    cli::cli_abort("To calculate standard errors, `keep_inbag` needs to be `TRUE`.")
   }
   
   # Get date